refactor(test): Split out more unit/integration tests
