+++ newrelic/resource_newrelic_plugins_alert_condition_test.go
@@ -1,133 +1,16 @@
-// +build integration
+// +build integration unit
+//
+// Test Helpers
+//
 
 package newrelic
 
 import (
 	"fmt"
-	"regexp"
-	"testing"
 
-	"github.com/hashicorp/terraform-plugin-sdk/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/terraform"
 )
-
-func TestAccNewRelicPluginsAlertCondition_Basic(t *testing.T) {
-	if !nrInternalAccount {
-		t.Skipf("New Relic internal testing account required")
-	}
-
-	rName := acctest.RandString(5)
-	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckNewRelicPluginsAlertConditionDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: testAccCheckNewRelicPluginsAlertConditionConfig(rName),
-				Check: resource.ComposeTestCheckFunc(
-					testAccCheckNewRelicPluginsAlertConditionExists("newrelic_plugins_alert_condition.foo"),
-					resource.TestCheckResourceAttr(
-						"newrelic_plugins_alert_condition.foo", "name", fmt.Sprintf("tf-test-%s", rName)),
-					resource.TestCheckResourceAttr(
-						"newrelic_plugins_alert_condition.foo", "enabled", "false"),
-					resource.TestCheckResourceAttr(
-						"newrelic_plugins_alert_condition.foo", "runbook_url", "https://foo.example.com"),
-					resource.TestCheckResourceAttr(
-						"newrelic_plugins_alert_condition.foo", "entities.#", "1"),
-					resource.TestCheckResourceAttr(
-						"newrelic_plugins_alert_condition.foo", "term.#", "1"),
-					resource.TestCheckResourceAttr(
-						"newrelic_plugins_alert_condition.foo", "term.1025554152.duration", "5"),
-					resource.TestCheckResourceAttr(
-						"newrelic_plugins_alert_condition.foo", "term.1025554152.operator", "below"),
-					resource.TestCheckResourceAttr(
-						"newrelic_plugins_alert_condition.foo", "term.1025554152.priority", "critical"),
-					resource.TestCheckResourceAttr(
-						"newrelic_plugins_alert_condition.foo", "term.1025554152.threshold", "0.75"),
-					resource.TestCheckResourceAttr(
-						"newrelic_plugins_alert_condition.foo", "term.1025554152.time_function", "all"),
-				),
-			},
-			{
-				Config: testAccCheckNewRelicPluginsAlertConditionConfigUpdated(rName),
-				Check: resource.ComposeTestCheckFunc(
-					testAccCheckNewRelicPluginsAlertConditionExists("newrelic_plugins_alert_condition.foo"),
-					resource.TestCheckResourceAttr(
-						"newrelic_plugins_alert_condition.foo", "name", fmt.Sprintf("tf-test-updated-%s", rName)),
-					resource.TestCheckResourceAttr(
-						"newrelic_plugins_alert_condition.foo", "enabled", "true"),
-					resource.TestCheckResourceAttr(
-						"newrelic_plugins_alert_condition.foo", "runbook_url", "https://bar.example.com"),
-					resource.TestCheckResourceAttr(
-						"newrelic_plugins_alert_condition.foo", "entities.#", "1"),
-					resource.TestCheckResourceAttr(
-						"newrelic_plugins_alert_condition.foo", "term.#", "1"),
-					resource.TestCheckResourceAttr(
-						"newrelic_plugins_alert_condition.foo", "term.1944209821.duration", "10"),
-					resource.TestCheckResourceAttr(
-						"newrelic_plugins_alert_condition.foo", "term.1944209821.operator", "below"),
-					resource.TestCheckResourceAttr(
-						"newrelic_plugins_alert_condition.foo", "term.1944209821.priority", "critical"),
-					resource.TestCheckResourceAttr(
-						"newrelic_plugins_alert_condition.foo", "term.1944209821.threshold", "0.65"),
-					resource.TestCheckResourceAttr(
-						"newrelic_plugins_alert_condition.foo", "term.1944209821.time_function", "all"),
-				),
-			},
-		},
-	})
-}
-
-func TestAccNewRelicPluginsAlertCondition(t *testing.T) {
-	if !nrInternalAccount {
-		t.Skipf("New Relic internal testing account required")
-	}
-
-	resourceName := "newrelic_plugins_alert_condition.foo"
-	rName := acctest.RandString(5)
-
-	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckNewRelicPluginsAlertConditionDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: testAccCheckNewRelicPluginsAlertConditionConfig(rName),
-			},
-
-			{
-				ResourceName:      resourceName,
-				ImportState:       true,
-				ImportStateVerify: true,
-			},
-		},
-	})
-}
-
-func TestAccNewRelicPluginsAlertCondition_MissingPolicy(t *testing.T) {
-	if !nrInternalAccount {
-		t.Skipf("New Relic internal testing account required")
-	}
-
-	rName := acctest.RandString(5)
-
-	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckNewRelicPluginsAlertConditionDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: testAccCheckNewRelicPluginsAlertConditionConfig(rName),
-			},
-			{
-				PreConfig: testAccDeleteNewRelicAlertPolicy(fmt.Sprintf("tf-test-%s", rName)),
-				Config:    testAccCheckNewRelicPluginsAlertConditionConfig(rName),
-				Check:     testAccCheckNewRelicPluginsAlertConditionExists("newrelic_plugins_alert_condition.foo"),
-			},
-		},
-	})
-}
 
 func testAccCheckNewRelicPluginsAlertConditionDestroy(s *terraform.State) error {
 	client := testAccProvider.Meta().(*ProviderConfig).NewClient
@@ -183,23 +66,6 @@ func testAccCheckNewRelicPluginsAlertConditionExists(n string) resource.TestChec
 
 		return nil
 	}
-}
-
-func TestAccNewRelicPluginsAlertCondition_NameGreaterThan64Char(t *testing.T) {
-	avoidEmptyAccountID()
-	expectedErrorMsg, _ := regexp.Compile(`expected length of name to be in the range \(1 \- 64\)`)
-	rName := acctest.RandString(5)
-	resource.ParallelTest(t, resource.TestCase{
-		IsUnitTest:   true,
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckNewRelicPluginsAlertConditionDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config:      testAccNewRelicPluginsAlertConditionNameGreaterThan64Char(rName),
-				ExpectError: expectedErrorMsg,
-			},
-		},
-	})
 }
 
 func testAccCheckNewRelicPluginsAlertConditionConfig(rName string) string {
@@ -319,22 +185,6 @@ resource "newrelic_plugins_alert_condition" "foo" {
 `, resourceName, testAccExpectedApplicationName)
 }
 
-func TestAccNewRelicPluginsAlertCondition_NameLessThan1Char(t *testing.T) {
-	avoidEmptyAccountID()
-	expectedErrorMsg, _ := regexp.Compile(`expected length of name to be in the range \(1 \- 64\)`)
-	resource.ParallelTest(t, resource.TestCase{
-		IsUnitTest:   true,
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckNewRelicPluginsAlertConditionDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config:      testAccNewRelicPluginsAlertConditionNameLessThan1Char(),
-				ExpectError: expectedErrorMsg,
-			},
-		},
-	})
-}
-
 func testAccNewRelicPluginsAlertConditionNameLessThan1Char() string {
 	return `
 provider "newrelic" {
@@ -374,22 +224,6 @@ resource "newrelic_plugins_alert_condition" "foo" {
 `
 }
 
-func TestAccNewRelicPluginsAlertCondition_TermDurationGreaterThan120(t *testing.T) {
-	avoidEmptyAccountID()
-	expectedErrorMsg, _ := regexp.Compile(`expected term.0.duration to be in the range \(5 - 120\)`)
-	resource.ParallelTest(t, resource.TestCase{
-		IsUnitTest:   true,
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckNewRelicPluginsAlertConditionDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config:      testAccNewRelicPluginsAlertConditionTermDurationGreaterThan120(),
-				ExpectError: expectedErrorMsg,
-			},
-		},
-	})
-}
-
 func testAccNewRelicPluginsAlertConditionTermDurationGreaterThan120() string {
 	return `
 provider "newrelic" {
@@ -427,22 +261,6 @@ resource "newrelic_plugins_alert_condition" "foo" {
 	}
 }
 `
-}
-
-func TestAccNewRelicPluginsAlertCondition_TermDurationLessThan5(t *testing.T) {
-	avoidEmptyAccountID()
-	expectedErrorMsg, _ := regexp.Compile(`expected term.0.duration to be in the range \(5 - 120\)`)
-	resource.ParallelTest(t, resource.TestCase{
-		IsUnitTest:   true,
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckNewRelicPluginsAlertConditionDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config:      testAccNewRelicPluginsAlertConditionTermDurationLessThan5(),
-				ExpectError: expectedErrorMsg,
-			},
-		},
-	})
 }
 
 func testAccNewRelicPluginsAlertConditionTermDurationLessThan5() string {
